# Conflicts:
#	shell/conf/node_minerva-101.conf

+++ aloja-web/Controller/MLCacheController.php
@@ -15,9 +15,9 @@ class MLCacheController extends AbstractController
 		$jsonLearners = '';
 		try
 		{
-            $dbml = new \PDO($this->container->get('config')['db_conn_chain'], $this->container->get('config')['mysql_user'], $this->container->get('config')['mysql_pwd']);
-            $dbml->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
-            $dbml->setAttribute(\PDO::ATTR_EMULATE_PREPARES, false);
+			$dbml = new \PDO($this->container->get('config')['db_conn_chain'], $this->container->get('config')['mysql_user'], $this->container->get('config')['mysql_pwd']);
+			$dbml->setAttribute(\PDO::ATTR_ERRMODE, \PDO::ERRMODE_EXCEPTION);
+			$dbml->setAttribute(\PDO::ATTR_EMULATE_PREPARES, false);
 
 
 			if (isset($_GET['ccache']))// && isset($_SERVER['HTTP_REFERER']) && $_SERVER['HTTP_REFERER'] != $cache_allow)
@@ -70,6 +70,9 @@ class MLCacheController extends AbstractController
 				$query = "DELETE FROM aloja_ml.minconfigs WHERE id_minconfigs='".$_GET['rmm']."'";
 				if ($dbml->query($query) === FALSE) throw new \Exception('Error when removing a minconfig from DB');
 
+				$query = "DELETE FROM aloja_ml.model_storage WHERE id_hash='".$_GET['rmm']."' AND type='minconf'";
+				if ($dbml->query($query) === FALSE) throw new \Exception('Error when removing a model from DB');
+
 				$command = 'rm -f '.getcwd().'/cache/query/'.$_GET['rmm'].'*';
 				$output[] = shell_exec($command);
  			}
@@ -106,7 +109,7 @@ class MLCacheController extends AbstractController
 				FROM (	SELECT s.*, COUNT(r.id_resolution) AS num_resolutions
 					FROM (	SELECT j.*, COUNT(m.id_minconfigs) AS num_minconfigs
 						FROM (	SELECT DISTINCT l.id_learner AS id_learner, l.algorithm AS algorithm,
-								l.creation_time AS creation_time, l.model AS model,
+								l.creation_time AS creation_time, l.model AS model, l.dataslice AS advanced,
 								COUNT(p.id_prediction) AS num_preds
 							FROM aloja_ml.learners AS l LEFT JOIN aloja_ml.predictions AS p ON l.id_learner = p.id_learner
 							GROUP BY l.id_learner
@@ -121,10 +124,10 @@ class MLCacheController extends AbstractController
 			$jsonLearners = '[';
 		    	foreach($rows as $row)
 			{
-				$jsonLearners = $jsonLearners.(($jsonLearners=='[')?'':',')."['".$row['id_learner']."','".$row['algorithm']."','".$row['model']."','".$row['creation_time']."','".$row['num_preds']."','".$row['num_minconfigs']."','".$row['num_resolutions']."','".$row['num_trees']."','<a href=\'/mlclearcache?rml=".$row['id_learner']."\'>Remove</a>']";
+				$jsonLearners = $jsonLearners.(($jsonLearners=='[')?'':',')."['".$row['id_learner']."','".$row['algorithm']."','".$row['model']."','".$row['advanced']."','".$row['creation_time']."','".$row['num_preds']."','".$row['num_minconfigs']."','".$row['num_resolutions']."','".$row['num_trees']."','<a href=\'/mlclearcache?rml=".$row['id_learner']."\'>Remove</a>']";
 			}
 			$jsonLearners = $jsonLearners.']';
-			$jsonLearningHeader = "[{'title':'ID'},{'title':'Algorithm'},{'title':'Model'},{'title':'Creation'},{'title':'Predictions'},{'title':'MinConfigs'},{'title':'Resolutions'},{'title':'Trees'},{'title':'Actions'}]";
+			$jsonLearningHeader = "[{'title':'ID'},{'title':'Algorithm'},{'title':'Model'},{'title':'Advanced'},{'title':'Creation'},{'title':'Predictions'},{'title':'MinConfigs'},{'title':'Resolutions'},{'title':'Trees'},{'title':'Actions'}]";
 
 			// Compilation of Minconfs on Cache
 			$query="SELECT mj.*, COUNT(mc.sid_minconfigs_centers) AS num_centers
@@ -160,20 +163,20 @@ class MLCacheController extends AbstractController
 			$jsonResolutionsHeader = "[{'title':'ID'},{'title':'Learner'},{'title':'Model'},{'title':'Creation'},{'title':'Sigma'},{'title':'Instances'},{'title':'Actions'}]";
 
 			// Compilation of Summaries on Cache
-			$query="SELECT DISTINCT id_summaries, model, creation_time
+			$query="SELECT DISTINCT id_summaries, model, dataslice, creation_time
 				FROM aloja_ml.summaries
 				";
 			$rows = $dbml->query($query);
 			$jsonSummaries = '[';
 		    	foreach($rows as $row)
 			{
-				$jsonSummaries = $jsonSummaries.(($jsonSummaries=='[')?'':',')."['".$row['id_summaries']."','".$row['model']."','".$row['creation_time']."','<a href=\'/mlclearcache?rms=".$row['id_summaries']."\'>Remove</a>']";
+				$jsonSummaries = $jsonSummaries.(($jsonSummaries=='[')?'':',')."['".$row['id_summaries']."','".$row['model']."','".$row['dataslice']."','".$row['creation_time']."','<a href=\'/mlclearcache?rms=".$row['id_summaries']."\'>Remove</a>']";
 			}
 			$jsonSummaries = $jsonSummaries.']';
-			$jsonSummariesHeader = "[{'title':'ID'},{'title':'Model'},{'title':'Creation'},{'title':'Actions'}]";
+			$jsonSummariesHeader = "[{'title':'ID'},{'title':'Model'},{'title':'Advanced'},{'title':'Creation'},{'title':'Actions'}]";
 
 			// Compilation of Precisions on Cache
-			$query="SELECT id_precision, model, creation_time
+			$query="SELECT id_precision, model, dataslice, creation_time
 				FROM aloja_ml.precisions
 				GROUP BY id_precision
 				";
@@ -181,10 +184,10 @@ class MLCacheController extends AbstractController
 			$jsonPrecisions = '[';
 		    	foreach($rows as $row)
 			{
-				$jsonPrecisions = $jsonPrecisions.(($jsonPrecisions=='[')?'':',')."['".$row['id_precision']."','".$row['model']."','".$row['creation_time']."','<a href=\'/mlclearcache?rmp=".$row['id_precision']."\'>Remove</a>']";
+				$jsonPrecisions = $jsonPrecisions.(($jsonPrecisions=='[')?'':',')."['".$row['id_precision']."','".$row['model']."','".$row['dataslice']."','".$row['creation_time']."','<a href=\'/mlclearcache?rmp=".$row['id_precision']."\'>Remove</a>']";
 			}
 			$jsonPrecisions = $jsonPrecisions.']';
-			$jsonPrecisionsHeader = "[{'title':'ID'},{'title':'Model'},{'title':'Creation'},{'title':'Actions'}]";
+			$jsonPrecisionsHeader = "[{'title':'ID'},{'title':'Model'},{'title':'Advanced'},{'title':'Creation'},{'title':'Actions'}]";
 
 			$dbml = null;
 		}
@@ -193,9 +196,9 @@ class MLCacheController extends AbstractController
 			$this->container->getTwig ()->addGlobal ( 'message', $e->getMessage () . "\n" );
 			$output = array();
 		}
-		echo $this->container->getTwig()->render('mltemplate/mlclearcache.html.twig',
+
+		return $this->render('mltemplate/mlclearcache.html.twig',
 			array(
-				'selected' => 'mlclearcache',
 				'learners' => $jsonLearners,
 				'header_learners' => $jsonLearningHeader,
 				'minconfs' => $jsonMinconfs,
